[Refactor] 중복 패키지 제거
